Changes added

+++ web-ems/src/app/incidentdashboard/incidentdashboard.component.ts
@@ -1,3 +1,5 @@
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { DialogComponent } from './../dialog/dialog.component';
 import { AfterViewInit, Component, OnInit, ViewChild, } from '@angular/core';
 import {IncidentService} from './incidentdashboard.service'
 import {incident} from './incidentdashboard.model'
@@ -5,6 +7,7 @@ import {MatTableDataSource} from '@angular/material/table';
 import {MatPaginator} from '@angular/material/paginator';
 import {MatSort} from '@angular/material/sort';
 import { LoginService } from '../login/login.service';
+import { MatDialog } from '@angular/material/dialog';
 
 
 @Component({
@@ -26,7 +29,7 @@ export class IncidentdashboardComponent implements OnInit,AfterViewInit {
   ngAfterViewInit() {
 
   }
-  constructor(private incidentService: IncidentService,private loginservice:LoginService) {}
+  constructor(private incidentService: IncidentService,private loginservice:LoginService,private dialog:MatDialog, private _snackBar:MatSnackBar) {}
   remove=false;
   currentrole:any;
 
@@ -57,13 +60,34 @@ export class IncidentdashboardComponent implements OnInit,AfterViewInit {
 
   delete(form_id:any){
 
-    this.incidentService.deleteUser(form_id).subscribe((result)=>{
-      //console.log(result);
-      this.ngOnInit();
-      alert("Deleted Successfully");
+    if(confirm("Are you sure you want to permanently delete this form?")==true){
+
+      this.incidentService.deleteUser(form_id).subscribe((result)=>{
+        //console.log(result);
+        this.ngOnInit();
+        this._snackBar.open('Deleted!','',{
+          verticalPosition:'top',
+         // horizontalPosition:'center',
+          panelClass:'edit'
+        })
+
+      })
+    }
+  }
+
+  /*
+  openDialog(){
+    this.dialog.open(DialogComponent,{
+      disableClose:true
     })
+    const form_id={
+
+    }
+
 
   }
+
+  */
 
   MenuDisplay(){
     if(this.loginservice.getToken()!='')
